cleaning code, adding test to review controller and refactoring reviewservice

+++ src/main/java/com/novatech/cybertech/services/implementation/ReviewManagementServiceImp.java
@@ -5,11 +5,16 @@ import com.novatech.cybertech.dto.request.review.ReviewCreateRequestDto;
 import com.novatech.cybertech.dto.request.review.ReviewUpdateRequestDto;
 import com.novatech.cybertech.dto.response.moderation.ModerationResponseDto;
 import com.novatech.cybertech.dto.response.review.ReviewResponseDto;
+import com.novatech.cybertech.entities.OrderEntity;
+import com.novatech.cybertech.entities.ProductEntity;
 import com.novatech.cybertech.entities.ReviewEntity;
-import com.novatech.cybertech.exceptions.CommentPostNotAllowedException;
-import com.novatech.cybertech.exceptions.ReviewNotFoundException;
+import com.novatech.cybertech.entities.UserEntity;
+import com.novatech.cybertech.exceptions.*;
 import com.novatech.cybertech.mappers.entity.ReviewMapper;
+import com.novatech.cybertech.repositories.OrderRepository;
+import com.novatech.cybertech.repositories.ProductRepository;
 import com.novatech.cybertech.repositories.ReviewRepository;
+import com.novatech.cybertech.repositories.UserRepository;
 import com.novatech.cybertech.services.core.ModerationService;
 import com.novatech.cybertech.services.core.ReviewManagementService;
 import lombok.RequiredArgsConstructor;
@@ -17,8 +22,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.UUID;
 
 import static com.novatech.cybertech.utils.DataGenerator.generateReviewEntity;
@@ -31,62 +34,83 @@ public class ReviewManagementServiceImp implements ReviewManagementService {
     private static final float HATEFUL_COMMENT_SCORE_THRESHOLD = 0.7f;
 
     private final ReviewMapper reviewMapper;
+    private final UserRepository userRepository;
+    private final OrderRepository orderRepository;
     private final ReviewRepository reviewRepository;
     private final ModerationService moderationService;
+    private final ProductRepository productRepository;
+
 
     @Override
     @Transactional(readOnly = true)
-    public Collection<ReviewResponseDto> getAll() {
-        return reviewMapper.mapFromEntityToResponseDto(reviewRepository.findAll());
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public ReviewResponseDto getByUUID(UUID uuid) {
+    public ReviewResponseDto getByUUID(final UUID uuid) {
         return reviewMapper.mapFromEntityToResponseDto(reviewRepository.findByUuid(uuid).orElseThrow(() -> new ReviewNotFoundException("No review with the UUID : " + uuid + " found")));
     }
 
-    @Override
-    @Transactional(readOnly = true)
-    public Collection<ReviewResponseDto> getByUUIDs(Collection<UUID> uuids) {
-        return reviewMapper.mapFromEntityToResponseDto(reviewRepository.findAllByUuidIn(uuids));
-    }
 
     @Override
     @Transactional
-    public ReviewResponseDto create(ReviewCreateRequestDto reviewCreateRequestDto) {
+    public ReviewResponseDto create(final ReviewCreateRequestDto reviewCreateRequestDto, final String keycloakId) {
+
+        final UserEntity user = userRepository.findByKeycloakIdAndIsActive(keycloakId, true).orElseThrow(() -> new UserDoesntExistsException("User that's trying to post this comment doesn't exists"));
+
+        final OrderEntity order = orderRepository.findByUuid(reviewCreateRequestDto.getOrderUuid()).orElseThrow(() -> new OrderNotFoundException("Order related to this review doesn't exists, order UUID : " + reviewCreateRequestDto.getOrderUuid()));
+
+        checkIfUserAlreadyBoughtThisProduct(reviewCreateRequestDto, keycloakId, order, user);
+
         ReviewEntity reviewEntity = reviewMapper.mapFromCreationRequestToEntity(reviewCreateRequestDto);
 
-        ModerationResponseDto moderationResponseDto = moderationService.checkIfIsHateful(reviewEntity.getComment());
+        final ProductEntity product = productRepository.findByUuid(reviewCreateRequestDto.getProductUuid()).orElseThrow(() -> new ProductNotFoundException("Product related to this review doesn't exists"));
 
-        if (moderationResponseDto.getScore() > HATEFUL_COMMENT_SCORE_THRESHOLD)
-            throw new CommentPostNotAllowedException("Your comment seems similar to other hateful comments detected on our website, our moderation team will review it and decide to post it or not.");
+        //I ned to recompile the moderation api module cause made some changes in it
+
+        ModerationResponseDto moderationResponseDto = moderationService.checkIfIsHateful(reviewEntity.getComment());
+        if (moderationResponseDto.getScore() > HATEFUL_COMMENT_SCORE_THRESHOLD) throw new CommentPostNotAllowedException("Your comment looks similar to other hateful comments detected on our website, our moderation team will review it and decide to post it or not.");
+
+        reviewEntity.setIsHateful(false);
+        reviewEntity.setUserEntity(user);
+        reviewEntity.setProductEntity(product);
 
         return reviewMapper.mapFromEntityToResponseDto(reviewRepository.save(reviewEntity));
     }
 
-    @Transactional
-    public Collection<ReviewResponseDto> createAutomatically(Collection<ReviewEntity> reviews) {
-        return new ArrayList<>(reviewMapper.mapFromEntityToResponseDto(reviews.stream().map(reviewRepository::save).toList()));
-    }
 
     @Override
     @Transactional
-    public ReviewResponseDto update(final ReviewUpdateRequestDto reviewCreateRequestDto) {
+    public ReviewResponseDto update(final ReviewUpdateRequestDto reviewCreateRequestDto, final String keycloakId) {
+
+        //TODO: Optimisation potentielle ici, chercher directement en base les users actifs
+        final boolean userExists = userRepository.existsByKeycloakIdAndIsActive(keycloakId, true);
+        if (!userExists) throw new UserDoesntExistsException("User that's trying to post this comment doesn't exists");
+
+
+        final ReviewEntity review = reviewRepository.findByUuid(reviewCreateRequestDto.getReviewUuid()).orElseThrow(() -> new ReviewNotFoundException("No review with the UUID : " + reviewCreateRequestDto.getReviewUuid() + " found"));
+
+        boolean isCurrentUserAuthorOfTheRequestReview = review.getUserEntity().getKeycloakId().equals(keycloakId);
+
+        if (!isCurrentUserAuthorOfTheRequestReview)
+            throw new ReviewNotFoundException("Current review Doesn't belongs to the connected user");
+
         return reviewMapper.mapFromEntityToResponseDto(reviewRepository.save(reviewMapper.mapFromUpdateRequestToEntity(reviewCreateRequestDto)));
     }
 
-    @Override
-    @Transactional
-    public void deleteByUUID(UUID uuid) {
-        reviewRepository.deleteByUuid(uuid);
-    }
 
     @Override
     @Transactional
-    public void deleteByUUIDs(Collection<UUID> uuids) {
-        reviewRepository.deleteAllByUuidIn(uuids);
+    public void deleteByUUID(final UUID uuid, final String keycloakId) {
+        final boolean userExists = userRepository.existsByKeycloakIdAndIsActive(keycloakId, true);
+        if (!userExists) throw new UserDoesntExistsException("User that's trying to post this comment doesn't exists");
+
+        final ReviewEntity review = reviewRepository.findByUuid(uuid).orElseThrow(() -> new ReviewNotFoundException("No review with the UUID : " + uuid + " found"));
+
+        boolean isCurrentUserAuthorOfTheRequestReview = review.getUserEntity().getKeycloakId().equals(keycloakId);
+
+        if (!isCurrentUserAuthorOfTheRequestReview)
+            throw new ReviewNotFoundException("Current review Doesn't belongs to the connected user");
+
+        reviewRepository.deleteByUuid(uuid);
     }
+
 
     @Transactional
     public ReviewResponseDto saveReview() {
@@ -103,5 +127,35 @@ public class ReviewManagementServiceImp implements ReviewManagementService {
         log.info("review saved");
         return reviewMapper.mapFromEntityToResponseDto(reviewRepository.save(reviewEntity));
 
+    }
+
+
+    private static void checkIfUserAlreadyBoughtThisProduct(ReviewCreateRequestDto reviewCreateRequestDto, String keycloakId, OrderEntity order, UserEntity user) {
+        final boolean orderContainsProduct = order.getOrderItemEntities().stream()
+                .map(orderItemEntity -> orderItemEntity.getProductEntity().getUuid())
+                .anyMatch(uuid -> reviewCreateRequestDto.getProductUuid().equals(uuid));
+
+        //TODO: maybe make this part more explicit in the future, can be confusing
+        if (!orderContainsProduct) {
+
+            log.warn("Product is not part of current Order (the order sent in the request DTO), a search will be done in all user's orders");
+
+            final boolean isProductPartOfUserOrders = user.getOrderEntities().stream()
+                    .flatMap(orderEntity -> orderEntity.getOrderItemEntities().stream())
+                    .map(orderItemEntity -> orderItemEntity.getProductEntity().getUuid())
+                    .anyMatch(uuid -> reviewCreateRequestDto.getProductUuid().equals(uuid));
+
+            if (!isProductPartOfUserOrders) {
+                log.error("The product on which the user {} is trying to post a comment on is not a part of his order, maybe the product has already been bought in another order", keycloakId);
+                throw new ProductNotFoundException("This product is not part of your order, you can put a review only on a product that you have already bought");
+            }
+        }
+    }
+
+    private static void isUserActive(final UserEntity user) {
+        if (!user.getIsActive()) {
+            log.error("User that's trying to post this comment is not active");
+            throw new UserNotActiveException("User that's trying to post this comment is not active");
+        }
     }
 }
